# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Store API Collect&Verything",
        "version": "0.1"
    },
    "paths": {
...
        "/api/stores/{id}/images": {
            "post": {
                "tags": [
                    "Stores"
                ],
                "summary": "Post a new store image",
                "description": "Create a new store image",
                "operationId": "addImage",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Store title",
                        "required": true
                    },
                    {
                        "name": "storeLine1",
                        "in": "query",
                        "description": "Store line 1",
                        "required": true
                    },
                    {
                        "name": "storeLine2",
                        "in": "query",
                        "description": "Store line 2"
                    },
                    {
                        "name": "zipCode",
                        "in": "query",
                        "description": "Zip code",
                        "required": true
                    },
                    {
                        "name": "city",
                        "in": "query",
                        "description": "City",
                        "required": true
                    },
                    {
                        "name": "country",
                        "in": "query",
                        "description": "Store country",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Account created"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/stores/{id}/images/{image_id}": {
            "delete": {
                "tags": [
                    "Stores"
                ],
                "summary": "Delete a store image",
                "description": "Soft delete a store image",
                "operationId": "removeImage",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Account id",
                        "required": true
                    },
                    {
                        "name": "image_id",
                        "in": "path",
                        "description": "Image id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Account deleted"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/stores/{id}/medias": {
            "post": {
                "tags": [
                    "Stores"
                ],
                "summary": "Post a new store media",
                "description": "Create a new store media",
                "operationId": "addMedia",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Store title",
                        "required": true
                    },
                    {
                        "name": "storeLine1",
                        "in": "query",
                        "description": "Store line 1",
                        "required": true
                    },
                    {
                        "name": "storeLine2",
                        "in": "query",
                        "description": "Store line 2"
                    },
                    {
                        "name": "zipCode",
                        "in": "query",
                        "description": "Zip code",
                        "required": true
                    },
                    {
                        "name": "city",
                        "in": "query",
                        "description": "City",
                        "required": true
                    },
                    {
                        "name": "country",
                        "in": "query",
                        "description": "Store country",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Account created"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/stores/{id}/medias/{mediaId}": {
            "delete": {
                "tags": [
                    "Stores"
                ],
                "summary": "Delete a store media",
                "description": "Soft delete a store media",
                "operationId": "removeMedia",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Account id",
                        "required": true
                    },
                    {
                        "name": "mediaId",
                        "in": "path",
                        "description": "Media id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Account deleted"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}